==> notebooks/00_exploration.ipynb <==
DATA_PATH=r'..\data\raw\survey_results_public.csv'

import numpy as np
import pandas as pd

raw_df=pd.read_csv(DATA_PATH)
raw_df.shape


raw_df.sample(1).iloc[0]


raw_df.info()

raw_df.describe()

questionaple_coulmns=['YearsCode','YearsCodePro','Age']
for col in questionaple_coulmns:
    print(col)
    print(raw_df[col].unique())
    print('-'*40)

raw_df.isnull().sum()

raw_df.describe()
