==> notebooks/01_preprocessing.ipynb <==
DATA_PATH=r'..\data\raw\survey_results_public.csv'

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import logging
import pickle 


REPLACE_DICT={
    'YearsCodePro':{'Less than 1 year':0,
                    'More than 50 years':51},
    'YearsCode':{'Less than 1 year':0,
                 'More than 50 years':51}               
}

raw_df=pd.read_csv(DATA_PATH)
df=raw_df.copy()

Preprocessing


#process yearscode and yearscodepro to be float instead of object
for coulmn,replacement in REPLACE_DICT.items():
    df[coulmn]=df[coulmn].replace(replacement).astype(np.float32)

df.describe()

def split_answers(data_series,delimeter=';'):
    """this fun take each coulmn name and return 
    the data frame which contain more than one value in string
    like list 
    """
    def issplittable(pd_series,delimeter):
        return pd_series.str.contains(delimeter)
    def split_answer(pd_series,delimeter):
        return pd_series.str.split(delimeter) 
    if not issplittable(data_series,delimeter).any():
        return data_series
    modified_series=split_answer(data_series,delimeter)
    #replace nulls with empty list 
    nulls=modified_series.isnull()
    modified_series.loc[nulls]=modified_series.loc[nulls].apply(lambda x:[])
    return modified_series


object_coulmns=df.select_dtypes('object').columns.to_list()
for coulmn in object_coulmns:
    df[coulmn]=split_answers(df[coulmn])


# df.sample(1).iloc[0]['YearsCode']
for i in REPLACE_DICT:
    print(raw_df.sample(1).iloc[0][i])
    print(df.sample(1).iloc[0][i])
    print("-"*40)

i=df.sample(1).index[0]
print(raw_df["LanguageHaveWorkedWith"].iloc[i])
print(df["LanguageHaveWorkedWith"].iloc[i])

df.sample(1).iloc[0]

EXPORT_PATH="dataframe.pkl"
df.to_pickle(EXPORT_PATH)
EXPORT_PATH

EXPORT_PATH
