==> notebooks/05_baseline_model.ipynb <==
ROLE_COLS=["DevType"]
TECH_COLS=['LanguageHaveWorkedWith',
       'DatabaseHaveWorkedWith',
       'PlatformHaveWorkedWith',
       'WebframeHaveWorkedWith',
       'MiscTechHaveWorkedWith',
       'ToolsTechHaveWorkedWith',
       'NEWCollabToolsHaveWorkedWith']

DATA_DIR=r'..\data\processed\2_cleaned_data.pkl'
MLFLOW_TRACKING_URI = '../models/mlruns/'
MLFLOW_EXPERIMENT_NAME = "skills_jobs_stackoverflow"

LOG_PATH = "../models/temp/"
LOG_DATA_PKL    =  "data.pkl"
LOG_MODEL_PKL   =  "model.pkl"
LOG_METRICS_PKL =  "metrics.pkl"


import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import train_test_split,GridSearchCV,cross_val_score
import os
from sklearn.linear_model import LogisticRegression
from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.multioutput import MultiOutputClassifier
from sklearn.metrics import precision_score,recall_score,f1_score,accuracy_score
from pathlib import Path
import mlflow 
from mlflow import MlflowClient

df=pd.read_pickle(DATA_DIR)
df

sampling


print(df["DevType"].sum(axis=0))

samples_per_class=1200
roles=df["DevType"].copy()

all_indices=[]
for col in roles.columns:
    # print(col)

    aa=roles[col].loc[roles[col]==1]
    if (roles[col].sum()>1200):
        aa=aa.sample(samples_per_class,replace=False,random_state=0)
    if(roles[col].sum()<1200):
        aa=aa.sample(samples_per_class,replace=True,random_state=0)
    # print(df["DevType"][col][aa.index])
    all_indices.append(list(aa.index))
    # print(aa.index.values)
all_indices=sum(all_indices,[])
new_df=df.loc[all_indices]
new_df

new_df["DevType"].sum(axis=0)

Path(MLFLOW_TRACKING_URI).mkdir(parents=True, exist_ok=True)
Path(LOG_PATH).mkdir(parents=True, exist_ok=True) 

# from mlflow import MlflowClient
# import mlflow
# # Initialize client and experiment
# mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
# client = MlflowClient()
# mlflow.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME)
# exp = client.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME)

X_train,X_test,y_train,y_test=train_test_split(new_df.drop("DevType",axis=1),new_df["DevType"],test_size=.3,random_state=52)


print(X_train.shape)
print(y_train.shape)

mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
client=MlflowClient()
mlflow.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME)
exp=client.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME)
if exp is None:
    print("Experiment not found ,creating new experiment")
    experiment_id=mlflow.create_experiment(MLFLOW_EXPERIMENT_NAME)
else:
    experiment_id=exp.experiment_id    

lr=make_pipeline(StandardScaler(),MultiOutputClassifier(LogisticRegression()))
lr.fit(X_train,y_train)

y_train_predicted=pd.DataFrame(lr.predict(X_train),columns=y_train.columns)
y_test_predicted=pd.DataFrame(lr.predict(X_test),columns=y_train.columns)
y_train_predicted

train_score={}
for col in y_train.columns:
    # print(precision_score(y_train[col],y_train_predicted[col]))
    train_score[col]=[(score(y_train[col],y_train_predicted[col])) for score in [precision_score,recall_score,f1_score,accuracy_score]]

train_score=pd.DataFrame(train_score,index=['precision_score','recall_score','f1_score','accuracy_score']).T
# print(train_score)
# train_score
mean_train_scores=train_score.mean()
print(mean_train_scores)
train_score

test_score={}
for col in y_test.columns:
    # print(precision_score(y_train[col],y_train_predicted[col]))
    test_score[col]=[(score(y_test[col],y_test_predicted[col])) for score in [precision_score,recall_score,f1_score,accuracy_score]]

test_score=pd.DataFrame(test_score,index=['precision_score','recall_score','f1_score','accuracy_score']).T

mean_test_scores=test_score.mean()
print(mean_test_scores)
test_score

LOG run

1_prepare

import pickle
data_details={'dara_path':DATA_DIR,
              'trainig_indices':X_train.index.tolist(),
              'testing_indices':X_test.index.tolist(),
              'feature_names':X_train.columns.droplevel(0).tolist(),
              'target_names':y_train.columns.tolist()
              }
with open(os.path.join(LOG_PATH, LOG_DATA_PKL), "wb") as output_file:
    pickle.dump(data_details, output_file)

model = {"model_description": "Baseline model: Logistic Regression ",
         "model_details": str(lr),
         "model_object": lr} 

with open(os.path.join(LOG_PATH, LOG_MODEL_PKL), "wb") as output_file:
    pickle.dump(model, output_file)

classes_metrics = {"train_scores": train_score, 
                   "test_scores":  test_score}

with open(os.path.join(LOG_PATH, LOG_METRICS_PKL), "wb") as output_file:
    pickle.dump(classes_metrics, output_file)

# import mlflow
with mlflow.start_run(experiment_id=exp.experiment_id, run_name=model["model_description"]):
    # Log pickles 
    mlflow.log_artifacts(LOG_PATH)
    
    # Track metrics 
    for metric, score in mean_test_scores.items():
        mlflow.log_metric(metric, score)
