==> notebooks/08_prediction.ipynb <==
MLFLOW_TRACK_URI=r'..\models\mlruns'
MLFLOW_RUN_ID='0f114561eea94a65b2b21f93b58af2d8'

LOG_PATH = "../models/temp/"
LOG_MODEL='model.pkl'
LOG_DATA='data.pkl'
LOG_METRICS='metrics.pkl'

FEATURE_CLUSTER_PATH='../data/processed/features_skills_clusters_description.yaml'

import pickle
import yaml 
import pandas
import mlflow
from mlflow import MlflowClient
import os
import pandas as pd

mlflow.set_tracking_uri(MLFLOW_TRACK_URI)
client=MlflowClient()

run=client.get_run(MLFLOW_RUN_ID)
artifacts_path=run.info.artifact_uri
artifacts_path
artifacts_path = artifacts_path.replace("file:///", "")  # Remove URI prefix
artifacts_path

# artifacts_path=r'file:///c:/Users/Mohamed Mosaad/Downloads/first_git/data_science_project/notebooks/../models/mlruns/113844087145227590/5252e0d0ca6a4ffa9e96fef11a250852/artifacts'



# "C:\Users\Mohamed Mosaad\Downloads\first_git\data_science_project\models\mlruns\113844087145227590\5252e0d0ca6a4ffa9e96fef11a250852\artifacts\model.pkl"

model


model_path=os.path.join(artifacts_path,LOG_MODEL)
with open (model_path,'rb') as aa:
    model=pickle.load(aa)
model

METRICS


metrics_path=os.path.join(artifacts_path,LOG_METRICS)
with open (metrics_path,'rb') as aa:
    metrics=pickle.load(aa)
metrics

DATA

data_path=os.path.join(artifacts_path,LOG_DATA)
with open (data_path,'rb') as aa:
    data=pickle.load(aa)
data

features=pd.Series(data["features_names"])
print(features)
targets=pd.Series(data["targets_names"])
targets

model

classifier=model["model_object"]

classifier

with open(FEATURE_CLUSTER_PATH,'r') as aa:
    clusters=yaml.safe_load(aa)
clusters    

aa={}
for cluster,skills in clusters.items():
    print(cluster,skills)
    for skill in skills:
        aa[skill]=cluster

molten_clusters = [(cluster_name, cluster_skill)
                   for cluster_name, cluster_skills in clusters.items()
                   for cluster_skill in cluster_skills]

clusters_df = pd.DataFrame(molten_clusters, columns=["cluster_name", "skill"])
clusters_df

group=['Python','SQL','Pandas','Keras','TensorFlow']
clusters_df["exist"]=clusters_df["skill"].isin(group)
clusters_df

cluster_features=clusters_df.groupby("cluster_name")['exist'].sum()


cluster_features

cluster_features.index


# skills_features=aa['exist']
# skills_features.index=aa["skills"]
# skills_features

skills_features=features[~features.isin(cluster_features.index)]
skills_features

group=['Python','SQL','Pandas','Keras','TensorFlow']

ohe_skills = pd.Series(skills_features.isin(group).astype(int).tolist(), 
                       index=skills_features)
ohe_skills


all_features=pd.concat([ohe_skills,cluster_features])
all_features

predictions = classifier.predict_proba([all_features.values])
predictions

positive_probs = [prob[0][1] for prob in predictions]
positive_probs

result=pd.Series(positive_probs,index=targets)
result
